requirements.txt
flask==3.1.0
flask-cors==5.0.1
yt-dlp>=2024.12.1
openai-whisper>=20231117
pydub>=0.25.1
gunicorn>=22.0.0
torch>=2.0.0
